// src/app/master/model/departmentmapping.ts
export class DepartmentMapping {
    companyname: string = "";
    departmentname: string = "";
    headname: string = "";
    status: boolean = false;
}

export class Companyselect {
    companyname: string = "";
    companycode: number = 0;
}

export class Departmentselect {
    deptname: string = "";
    deptcode: number = 0;
}
export class Roundselect {
    roundname: string = "";
    roundcode: number = 0;
}
export class Employeeselect {
    employeename: string = "";
    employeecode: number = 0;
}
export class Headselect {
    headname: string = "";
    headcode: number = 0;
}